AWS SQS Notification
This example shows SQS notification. After reading notification from sqs queue, notificationReceiver sends to to "critical" or "normal" SNS topics.
If it is sent to "critical" user has option to get email subscriber and lamda subscriber is always present. lambda subscriber logs that in cloudwatch log.

If it is sent to "normal" than lambda subscriber logs that in cloudwatch log.